Should be working much better now

+++ testing/js/spotifyAPI.js
@@ -1,30 +1,12 @@
-let accessToken = localStorage.getItem('spotifyAccessToken');
-
 function getSpotifyAuthToken() {
   const urlParams = new URLSearchParams(window.location.search);
   const authToken = urlParams.get('code');
   return authToken;
 }
 
-if (!accessToken) {
-  fetchSpotifyToken().then(tokens => {
-    accessToken = tokens.accessToken;
-    localStorage.setItem('spotifyAccessTokenExpiry', Date.now() + 45 * 60 * 1000);  // Expiry in 45 minutes
-    fetchCurrentlyPlaying(); 
-    setInterval(fetchCurrentlyPlaying, 5000);
-  });
-} else if (Date.now() > localStorage.getItem('spotifyAccessTokenExpiry')) {
-  // Token expired, fetch a new one
-  fetchSpotifyToken().then(tokens => {
-    accessToken = tokens.accessToken;
-    localStorage.setItem('spotifyAccessTokenExpiry', Date.now() + 45 * 60 * 1000);  // Expiry in 45 minutes
-  });
-} else {
-  fetchCurrentlyPlaying(); 
-  setInterval(fetchCurrentlyPlaying, 5000);
-}
-
 async function fetchSpotifyToken() {
+
+  
     const options = {
       method: 'POST',
       headers: {
@@ -41,22 +23,24 @@ async function fetchSpotifyToken() {
     try {
       const response = await fetch('https://accounts.spotify.com/api/token', options);
       const data = await response.json();
-        localStorage.setItem('spotifyRefreshToken', data.refresh_token);
+  
+      if (data.access_token && data.refresh_token) { // check if tokens exist in the response
         localStorage.setItem('spotifyAccessToken', data.access_token);
-        console.log(localStorage.getItem('spotifyAccessToken'));
-
+        localStorage.setItem('spotifyRefreshToken', data.refresh_token);
+      }
+  
       return { accessToken: data.access_token, refreshToken: data.refresh_token };
-
     } catch (err) {
       console.error(err);
     }
+
 }
 
 async function fetchCurrentlyPlaying() {
   const options = {
     method: 'GET',
     headers: {
-      Authorization: 'Bearer ' + accessToken
+      Authorization: 'Bearer ' + localStorage.getItem('spotifyAccessToken')
     }
   };
 
@@ -65,8 +49,53 @@ async function fetchCurrentlyPlaying() {
     const responseData = await response.json();
     const currentlyPlayingData = responseData;
     console.log(currentlyPlayingData.item.name);
+    console.log(currentlyPlayingData.item.artists[0].name);
+    console.log(currentlyPlayingData.item);
+    console.log(currentlyPlayingData.item.album.images[0].url);
     document.getElementById("nowPlaying").innerHTML = currentlyPlayingData.item.name;
+    document.getElementById("artistPlaying").innerHTML = currentlyPlayingData.item.artists[0].name;
+    document.getElementById("nowPlayingAlbum").src = currentlyPlayingData.item.album.images[0].url;
   } catch (err) {
     console.error(err);
   }
 }
+
+let localAccessToken = localStorage.getItem('spotifyAccessToken');
+
+if (localAccessToken){
+  console.log('Confirm IS token');
+} else if (!localAccessToken){
+  console.log('Confirm NO token');
+}
+
+
+  if (typeof localStorage.getItem('spotifyAccessToken') !== 'undefined'){
+    fetchSpotifyToken();
+    fetchCurrentlyPlaying(); 
+  }
+
+
+  // if (!localAccessToken) {
+  //   fetchSpotifyToken().then(tokens => {
+  //     accessToken = tokens.accessToken;
+  //     localStorage.setItem('spotifyAccessTokenExpiry', Date.now() + 45 * 60 * 1000);  // Expiry in 45 minutes
+  //     fetchCurrentlyPlaying(); 
+  //     setInterval(fetchCurrentlyPlaying, 5000);
+  //     console.log('There is no Access Token');
+  //     console.log(accessToken);
+  //   });
+  // } else if (Date.now() > localStorage.getItem('spotifyAccessTokenExpiry')) {
+  //   // Token expired, fetch a new one
+  //   fetchSpotifyToken().then(tokens => {
+  //     accessToken = tokens.accessToken;
+  //     localStorage.setItem('spotifyAccessTokenExpiry', Date.now() + 45 * 60 * 1000);  // Expiry in 45 minutes
+  //   });
+  // } else {
+  //   fetchSpotifyToken().then(tokens => {
+  //     accessToken = tokens.accessToken;  // Update the accessToken after fetching
+  //     fetchCurrentlyPlaying(); 
+  //     setInterval(fetchCurrentlyPlaying, 5000);
+  //     console.log('There is an Access Token');
+  //     console.log(accessToken);
+  //   });
+  // }
